Removed (almost) unused scripts correlation.py and jacobian.py
